refactored out response handlers
