+++ lib/event_listeners/foods.js
@@ -1,5 +1,6 @@
 const $ = require('jquery')
-import {filterFoods, handleError, createNewFood, updateFood, deleteFood, sortByCaloriesDesc, sortByCaloriesAsc, sortByCaloriesOrig} from '../ajax_requests/food'
+import {createNewFood, updateFood, deleteFood} from '../ajax_requests/food'
+import {filterFoods, handleError, sortByCaloriesDesc, sortByCaloriesAsc, sortByCaloriesOrig} from '../response-handlers/food-responses'
 
 export function eventListenerFunction() {
   $('#search-foods').on('keyup', filterFoods);
